Use API v5 to get items
It seems that an update has been made on
the backend, and API v1 does not work anymore.

This commit is enough to be able to retrieve data again.
I am not able to tell if the data format is still the same
or if changes have to be done on clients.